CMake: Split auto pilot plugins
This also modify how a few headers are included - using the full
patch instead of just the file name. I understand that this is
different from what the software is doing till now, but this
improves readability. Having the full set of headers without the
folders is almost the same as having all the sources in the same
folder: hard to figure out where something came from and hard
to reason about the project structure.