CMake: Split AutopilotPlugins into different libraries
There's an error when I actually enable the code to
split the AutoPilotPlugins into many libraries, so the
code that does this is commented. It's a start, tougth.
Perhaps this should be a dynamic library instead of a static one?

The qmake version had no static libraries and everything was
compiled and linked at once, this can also be a way to test.

Besides that,  many headers now have the full path of the file,
this helps us to better understand the Architecture of the program